remove the Range method in the streamsMap
streamsMap.Range was only used to pass the stream flow control window to
all open streams after receiving the peer's transport parameters. Since
the session is only returned after the connection is forward-secure, it
is not possible to open any streams (except the crypto stream) before
receiving the transport parameters, so streamsMap.Range is not needed.